requirements.txt
PyQt5
Pillow
requests
numpy